refactor: update remove product method with additional validations and reorganize packages in config
