+++ app/src/main/java/com/marvim/wishlist/config/handler/exception/ProductAlreadyInWishlistException.java
@@ -1,4 +1,4 @@
-package com.marvim.wishlist.config.exception;
+package com.marvim.wishlist.config.handler.exception;
 
 public class ProductAlreadyInWishlistException extends RuntimeException {
     public ProductAlreadyInWishlistException(String clientId, String productId) {
