+++ app/src/main/java/com/marvim/wishlist/config/handler/exception/ProductNotFoundException.java
@@ -1,0 +1,7 @@
+package com.marvim.wishlist.config.handler.exception;
+
+public class ProductNotFoundException extends RuntimeException {
+    public ProductNotFoundException(String clientId, String productId) {
+        super(String.format("Product %s not found in customer wishlist with id: %s",productId, clientId));
+    }
+}
