+++ app/src/main/java/com/marvim/wishlist/config/handler/exception/WishlistNotFoundException.java
@@ -1,4 +1,4 @@
-package com.marvim.wishlist.config.exception;
+package com.marvim.wishlist.config.handler.exception;
 
 public class WishlistNotFoundException extends RuntimeException {
     public WishlistNotFoundException(String clientId) {
